preprocess.py
import cv2
import numpy as np

# Preprocess the input image
def preprocess_image(image_path, target_size=(224, 224)):
    image = cv2.imread(image_path)
    image = cv2.resize(image, target_size)
    image = image / 255.0
    image = np.expand_dims(image, axis=0)  # Add batch dimension
    return image
