requirements.txt
Flask==2.2.5
numpy==1.23.5
tensorflow==2.13.0
Pillow==9.4.0
gunicorn==20.1.0
scikit-learn==1.2.2
